requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
amqp==5.1.1
    # via kombu
asgiref==3.7.2
    # via django
async-timeout==4.0.2
    # via redis
billiard==4.1.0
    # via celery
celery[redis,yaml]==5.3.1
    # via -r requirements.in
click==8.1.5
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
django==4.2.3
    # via
    #   -r requirements.in
    #   django-filter
    #   django-redis
    #   djangorestframework
django-environ==0.10.0
    # via -r requirements.in
django-filter==23.2
    # via -r requirements.in
django-redis==5.3.0
    # via -r requirements.in
djangorestframework==3.14.0
    # via -r requirements.in
kombu==5.3.1
    # via celery
prompt-toolkit==3.0.39
    # via click-repl
psycopg2-binary==2.9.6
    # via -r requirements.in
pyjwt==2.7.0
    # via -r requirements.in
python-dateutil==2.8.2
    # via celery
pytz==2023.3
    # via djangorestframework
pyyaml==6.0
    # via celery
redis==4.6.0
    # via
    #   celery
    #   django-redis
six==1.16.0
    # via python-dateutil
sqlparse==0.4.4
    # via django
tzdata==2023.3
    # via celery
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.6
    # via prompt-toolkit
